+++ plutarch-test/src/Plutarch/PlutusTypeSpec.hs
@@ -4,10 +4,22 @@ module Plutarch.PlutusTypeSpec (spec) where
 
 import Test.Syd
 
-import Plutarch
+import Plutarch (ClosedTerm, pcon', pmatch')
+import Plutarch.Api.V1 (
+  PAddress (PAddress),
+  PCredential (PPubKeyCredential, PScriptCredential),
+  PScriptPurpose (PCertifying, PMinting, PRewarding, PSpending),
+ )
+import Plutarch.Builtin (pasByteStr, pasConstr)
 import Plutarch.Prelude
-
 import Plutarch.Test
+import Plutus.V1.Ledger.Address (Address (Address))
+import Plutus.V1.Ledger.Api (DCert (DCertGenesis), toData)
+import Plutus.V1.Ledger.Contexts (ScriptPurpose (Certifying, Minting, Rewarding, Spending), TxOutRef (TxOutRef))
+import Plutus.V1.Ledger.Credential (
+  Credential (PubKeyCredential, ScriptCredential),
+  StakingCredential (StakingPtr),
+ )
 
 spec :: Spec
 spec = do
@@ -18,10 +30,196 @@ spec = do
       "swap" @\ do
         "A" @| swap (pcon A) @== pcon B
         "B" @| swap (pcon B) @== pcon A
+      "scottenc" @\ do
+        "PMaybe"
+          @| ( let a = 42 :: Term s PInteger
+                in pmatch (pcon $ PJust a) $ \case
+                    PJust x -> x
+                    -- We expect this perror not to be evaluated eagerly when mx
+                    -- is a PJust.
+                    PNothing -> perror
+             )
+        "PPair"
+          @| ( let a = 42 :: Term s PInteger
+                   b = "Universe" :: Term s PString
+                in pmatch (pcon (PPair a b) :: Term s (PPair PInteger PString)) $ \(PPair _ y) -> y
+             )
     describe "instances-sanity" $ do
       plutarchDevFlagDescribe $ do
         it "PBuiltinList" $ do
           pmatchTargetEval $ pconstant [1 :: Integer, 2, 3, 4]
+    deconstrSpec
+
+{- | For comparing typed and untyped data deconstruction approaches.
+
+We ideally want the typed and raw versions to have as little deviation as possible.
+-}
+deconstrSpec :: Spec
+deconstrSpec = do
+  describe "deconstr" . pgoldenSpec $ do
+    "matching" @\ do
+      "typed" @\ do
+        "newtype"
+          @| plam
+            ( \x -> pmatch x $ \(PAddress addrFields) ->
+                addrFields
+            )
+            # pconstant addrPC
+        "sumtype(ignore-fields)"
+          @| plam
+            ( \x -> pmatch x $ \case
+                PMinting _ -> pconstant ()
+                _ -> perror
+            )
+            # pconstant minting
+        "sumtype(partial-match)"
+          @| plam
+            ( \x -> pmatch x $ \case
+                PMinting hs -> hs
+                _ -> perror
+            )
+            # pconstant minting
+        "sumtype(exhaustive)" @\ do
+          benchPurpose $
+            plam
+              ( \x -> pmatch x $ \case
+                  PMinting f -> plet f $ const $ phexByteStr "01"
+                  PSpending f -> plet f $ const $ phexByteStr "02"
+                  PRewarding f -> plet f $ const $ phexByteStr "03"
+                  PCertifying f -> plet f $ const $ phexByteStr "04"
+              )
+        "sumtype(exhaustive)(ignore-fields)" @\ do
+          benchPurpose $
+            plam
+              ( \x -> pmatch x $ \case
+                  PMinting _ -> phexByteStr "01"
+                  PSpending _ -> phexByteStr "02"
+                  PRewarding _ -> phexByteStr "03"
+                  PCertifying _ -> phexByteStr "04"
+              )
+      "raw" @\ do
+        "newtype"
+          @| plam
+            ( \x ->
+                psndBuiltin #$ pasConstr # x
+            )
+            #$ pconstant
+              (toData addrPC)
+        "sumtype(ignore-fields)"
+          @| plam
+            ( \x ->
+                pif
+                  ((pfstBuiltin #$ pasConstr # x) #== 0)
+                  (pconstant ())
+                  perror
+            )
+            #$ pconstant (toData minting)
+        "sumtype(partial-match)"
+          @| plam
+            ( \x ->
+                plet (pasConstr # x) $ \d ->
+                  pif
+                    (pfstBuiltin # d #== 0)
+                    (psndBuiltin # d)
+                    perror
+            )
+            #$ pconstant (toData minting)
+        "sumtype(exhaustive)" @\ do
+          benchPurpose' $
+            plam
+              ( \x ->
+                  plet (pasConstr # x) $ \d ->
+                    plet (pfstBuiltin # d) $ \constr ->
+                      plet (psndBuiltin # d) $ \_ ->
+                        pif
+                          (constr #== 1)
+                          (phexByteStr "02")
+                          $ pif
+                            (constr #== 2)
+                            (phexByteStr "03")
+                            $ pif
+                              (constr #== 3)
+                              (phexByteStr "04")
+                              $ phexByteStr "01"
+              )
+        "sumtype(exhaustive)(ignore-fields)" @\ do
+          benchPurpose' $
+            plam
+              ( \x -> do
+                  plet (pfstBuiltin #$ pasConstr # x) $ \constr ->
+                    pif
+                      (constr #== 1)
+                      (phexByteStr "02")
+                      $ pif
+                        (constr #== 2)
+                        (phexByteStr "03")
+                        $ pif
+                          (constr #== 3)
+                          (phexByteStr "04")
+                          $ phexByteStr "01"
+              )
+    "fields" @\ do
+      "typed" @\ do
+        "extract-single"
+          @| plam
+            ( \x ->
+                pfield @"credential" # x
+            )
+            # pconstant addrSC
+      "raw" @\ do
+        "extract-single"
+          @| plam
+            ( \x ->
+                phead #$ psndBuiltin #$ pasConstr # x
+            )
+            #$ pconstant
+          $ toData addrSC
+    "combined" @\ do
+      "typed" @\ do
+        "toValidatorHash"
+          @| plam
+            ( \x ->
+                pmatch (pfromData $ pfield @"credential" # x) $ \case
+                  PPubKeyCredential _ ->
+                    pcon PNothing
+                  PScriptCredential credFields ->
+                    pcon . PJust $ pto $ pfromData $ pfield @"_0" # credFields
+            )
+            # pconstant addrSC
+      "raw" @\ do
+        "toValidatorHash"
+          @| plam
+            ( \x ->
+                let cred = phead #$ psndBuiltin #$ pasConstr # x
+                 in plet (pasConstr # cred) $ \deconstrCred ->
+                      pif
+                        (pfstBuiltin # deconstrCred #== 0)
+                        (pcon PNothing)
+                        $ pcon . PJust $ pasByteStr #$ phead #$ psndBuiltin # deconstrCred
+            )
+            # pconstant (toData addrSC)
+  where
+    addrSC = Address (ScriptCredential "ab") Nothing
+    addrPC = Address (PubKeyCredential "ab") Nothing
+    minting :: ScriptPurpose
+    minting = Minting ""
+    spending = Spending (TxOutRef "ab" 0)
+    rewarding = Rewarding (StakingPtr 42 0 7)
+    certifying = Certifying DCertGenesis
+    -- Bench given function feeding in all 4 types of script purpose (typed).
+    benchPurpose :: ClosedTerm (PScriptPurpose :--> PByteString) -> PlutarchGoldens
+    benchPurpose f = do
+      "minting" @| f # pconstant minting
+      "spending" @| f # pconstant spending
+      "rewarding" @| f # pconstant rewarding
+      "certifying" @| f # pconstant certifying
+    -- Bench given function feeding in all 4 types of script purpose (untyped).
+    benchPurpose' :: ClosedTerm (PData :--> PByteString) -> PlutarchGoldens
+    benchPurpose' f = do
+      "minting" @| f #$ pconstant $ toData minting
+      "spending" @| f #$ pconstant $ toData spending
+      "rewarding" @| f #$ pconstant $ toData rewarding
+      "certifying" @| f #$ pconstant $ toData certifying
 
 -- | Make sure the target of 'pmatch' is only evaluated once.
 pmatchTargetEval :: PlutusType p => ClosedTerm p -> Expectation